🗺️ Atlas: [architectural change] Splitting Checkpoint Blob
Extracted tests from src/storage/checkpoint.rs into src/storage/checkpoint/tests.rs, keeping the core logic in src/storage/checkpoint/mod.rs.
